refactor: LottoResult를 LottoTickets가 WinningNumbers를 받아 생성토록 변경
