+++ src/main/java/domain/LottoTickets.java
@@ -1,10 +1,10 @@
 package domain;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
+import java.util.*;
 
 public class LottoTickets {
+    private static final Integer INITIAL_VALUE_ONE = 1;
+
     private final List<LottoTicket> lottoTickets;
 
     public LottoTickets() {
@@ -25,6 +25,24 @@ public class LottoTickets {
 
     public void addAll(LottoTickets lottoTickets) {
         this.lottoTickets.addAll(lottoTickets.lottoTickets());
+    }
+
+    public LottoResult generateLottoResult(WinningNumbers winningNumbers)  {
+        Map<Rank, Integer> resultStatisticsMap = new HashMap<>();
+        for (LottoTicket lottoTicket : this.lottoTickets) {
+            Rank rank = winningNumbers.calculateRank(lottoTicket);
+            addCount(resultStatisticsMap, rank);
+        }
+        return new LottoResult(resultStatisticsMap);
+    }
+
+    private static void addCount(Map<Rank, Integer> resultStatisticsMap, Rank rank) {
+        if (resultStatisticsMap.containsKey(rank)) {
+            int count = resultStatisticsMap.get(rank);
+            resultStatisticsMap.put(rank, count + 1);
+            return;
+        }
+        resultStatisticsMap.put(rank, INITIAL_VALUE_ONE);
     }
 
     public int size() {
