+++ src/test/java/domain/LottoTicketsTest.java
@@ -4,9 +4,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
+import java.util.*;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -52,5 +50,29 @@ public class LottoTicketsTest {
         lottoTickets.addAll(lottoTicketsToBeAppended);
 
         assertThat(lottoTickets.size()).isEqualTo(4);
+    }
+
+    @DisplayName("일치하는 번호 개수에 따른 통계값을 구해 LottoResult로 리턴한.")
+    @Test
+    public void generateLottoResultTest() {
+        WinningNumbers winningNumbers = WinningNumbers.of(
+                LottoTicket.valueOf(Arrays.asList(1, 2, 3, 4, 5, 6)),
+                LottoNumber.from(7));
+        LottoTickets lottoTickets = LottoTickets.of(Arrays.asList(
+                LottoTicket.valueOf(Arrays.asList(1, 2, 3, 4, 5, 6)),
+                LottoTicket.valueOf(Arrays.asList(1, 2, 3, 4, 5, 7)),
+                LottoTicket.valueOf(Arrays.asList(1, 2, 3, 4, 6, 7)),
+                LottoTicket.valueOf(Arrays.asList(9, 10, 11, 12, 13, 14))
+        ));
+
+        LottoResult lottoResult = lottoTickets.generateLottoResult(winningNumbers);
+        Map<Rank, Integer> lottoStatisticsMap = lottoResult.result();
+        assertThat(lottoStatisticsMap).isEqualTo(new HashMap<Rank, Integer>() {
+            {
+                put(Rank.FIRST, 1);
+                put(Rank.SECOND, 2);
+                put(Rank.NOTHING, 1);
+            }
+        });
     }
 }
